Multitude of changes to asset & collection forms
toasts are now partially implemented, shown on errors. If user doesn't complete the collection info they cannot move on, if the user starts to edit an asset the app won't allow them to move on until they save the changes. Work on submission, next. When the user submits the collection it takes two clicks to make it work, maybe i'll have a popup appear on screen or somehthing housing another submission button.